+++ lib/ui/uikit/src/components/PlSplash/PlSplash.vue
@@ -4,20 +4,29 @@ import { PlLoaderCircular } from '@/components/PlLoaderCircular';
 
 withDefaults(defineProps<{
   /**
-   * Optional string that sets size for PlLoaderCircular.
-   */
-  size?: '16' | '24' | '48' | string;
-  /**
    * Optional string that sets text that will be shown below the PlLoaderCircular.
    */
-  text?: string;
-
-}>(), { size: '48', text: 'Loading' });
+  loadingText?: string;
+  /**
+   * If `true` the loading overlay is shown.
+   */
+  loading?: boolean;
+  /**
+   * @TODO
+   */
+  type?: 'table' | 'transparent';
+}>(), { loadingText: undefined, loading: false, type: undefined });
 </script>
 
 <template>
-  <div :class="style.container">
-    <PlLoaderCircular :size="size" />
-    <h3 :class="style.text">{{ text }}</h3>
+  <div :class="[style.splash, {[style.table]: type === 'table', [style.transparent]: type === 'transparent'}]">
+    <div v-if="loading" :class="[style.overlay]">
+      <div>
+        <!-- @TODO refactor PlLoaderCircular size property -->
+        <PlLoaderCircular size="48" />
+        <div v-if="loadingText" :class="style.text">{{ loadingText }}</div>
+      </div>
+    </div>
+    <slot />
   </div>
 </template>