+++ sdk/eslint-config/index.d.ts
@@ -1,0 +1,9 @@
+import type { Linter } from 'eslint'
+
+declare const model: Linter.Config[]
+declare const ui: Linter.Config[]
+
+export {
+  model,
+  ui,
+}