+++ sdk/ui-vue/src/lib.ts
@@ -7,6 +7,8 @@ import ValueOrErrorsComponent from './components/ValueOrErrorsComponent.vue';
 
 export { BlockLayout, PlAgDataTable, PlAgOverlayLoading, PlAgOverlayNoRows, ValueOrErrorsComponent };
 
+export * from './AgGridVue';
+
 export * from './components/PlAgColumnHeader';
 
 export * from './components/PlAgCellFile';
@@ -15,10 +17,7 @@ export * from './components/PlAgCellStatusTag';
 export * from './components/PlAgChartStackedBarCell';
 export * from './components/PlAgChartHistogramCell';
 
-export * from './components/PlAgDataTable/types';
-export * from './components/PlAgDataTable/sources/row-number';
-export * from './components/PlAgDataTable/sources/focus-row';
-export * from './components/PlAgDataTable/sources/menu-items';
+export * from './components/PlAgDataTable';
 
 export * from './components/PlAgCsvExporter';