+++ etc/blocks/download-file/block/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @milaboratories/milaboratories.test-download-file
 
+## 1.0.79
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-download-file.ui@1.0.79
+
+## 1.0.78
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-download-file.ui@1.0.78
+
+## 1.0.77
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-download-file.ui@1.0.77
+
+## 1.0.76
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-download-file.ui@1.0.76
+
 ## 1.0.75
 
 ### Patch Changes