+++ etc/blocks/enter-numbers/block/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @milaboratories/milaboratories.test-enter-numbers
 
+## 1.0.201
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.201
+
+## 1.0.200
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.200
+
+## 1.0.199
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.199
+
+## 1.0.198
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.198
+
 ## 1.0.197
 
 ### Patch Changes