+++ etc/blocks/model-test/block/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @milaboratories/milaboratories.test-block-model
 
+## 1.0.119
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-block-model.ui@1.0.119
+
+## 1.0.118
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-block-model.ui@1.0.118
+
+## 1.0.117
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-block-model.ui@1.0.117
+
+## 1.0.116
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-block-model.ui@1.0.116
+
 ## 1.0.115
 
 ### Patch Changes