+++ etc/blocks/pool-explorer/block/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @milaboratories/milaboratories.pool-explorer
 
+## 1.0.35
+
+### Patch Changes
+
+- @milaboratories/milaboratories.pool-explorer.ui@1.1.17
+
+## 1.0.34
+
+### Patch Changes
+
+- @milaboratories/milaboratories.pool-explorer.ui@1.1.16
+
+## 1.0.33
+
+### Patch Changes
+
+- @milaboratories/milaboratories.pool-explorer.ui@1.1.15
+
+## 1.0.32
+
+### Patch Changes
+
+- @milaboratories/milaboratories.pool-explorer.ui@1.1.14
+
 ## 1.0.31
 
 ### Patch Changes