+++ etc/uikit-playground/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @milaboratories/uikit-playground
 
+## 0.1.40
+
+### Patch Changes
+
+- Updated dependencies [0c5753b]
+  - @milaboratories/uikit@2.2.57
+
+## 0.1.39
+
+### Patch Changes
+
+- Updated dependencies [4f23447]
+  - @milaboratories/uikit@2.2.56
+
+## 0.1.38
+
+### Patch Changes
+
+- Updated dependencies [852dcbf]
+  - @milaboratories/uikit@2.2.55
+
 ## 0.1.37
 
 ### Patch Changes