+++ tests/workflow-tengo/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @platforma-sdk/workflow-tengo-tests
 
+## 1.5.0
+
+### Minor Changes
+
+- 205c17e: Added ability to filter only specific partitions on xsv export with `{partitions: {idx: [keys]}}` option.
+
+## 1.4.3
+
+### Patch Changes
+
+- fd653da: Run tests in parallel (max 3 files at a time)
+
 ## 1.4.2
 
 ### Patch Changes